+++ tuning_fork_samples/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.5
+
+* Version brought level with the rest of the family; nothing else changed.
+
 ## 0.1.4
 
 * Version brought level with the rest of the family; nothing else changed.